[feat] Enhancements for introspection and debugging
## Testing ##

This is a preview of a proposed change to the __init__() of classes in dis.py.
Currently it is difficult to create an instance of the classes with
predecided values, as their __init__() takes no arguments.
I propose to set object attributes from params in the __init__(), with
default values.