[refactor] use class attributes for fixed recordType and recordLength
These are moved from the __init__() to class attributes.
This pattern makes it less likely for users to accidentally initialise the class with the
wrong recordType or recordLength, where these values are known beforehand.

Class attributes are accessible by instances of the class, so the serialize() and parse()
methods are unaffected.

### Further changes

- added type annotations to attributes, marking them as being an enum, int, or float,
  along with their size in bits
- fixed init types for some attributes
- removed some unnecessary comments
- added table UID in comments for some enums